HasPririty and Index Table featured

+++ database/migrations/Auth/2020_10_07_225358_create_permission_groups_table.php
@@ -17,7 +17,7 @@ class CreatePermissionGroupsTable extends Migration
             $table->id();
             $table->boolean('active')->default(1);
             $table->string('group')->unique();
-            $table->integer('priority')->default(1);
+            $table->unsignedBigInteger('priority')->default(1);
             $table->timestamps();
             $table->foreignId('created_by')->nullable()->references('id')->on('users')->onDelete('set null');
             $table->foreignId('updated_by')->nullable()->references('id')->on('users')->onDelete('set null');
